move to util CallbackBytesIO
